refactor(client): :truck: move BaseModal
