+++ client/src/components/Pages/dispositif/Edition/CustomNavbar/CustomNavbarEdit/PublishModal/PublishModal.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useState } from "react";
 import { DeepPartialSkipArrayKey, useWatch } from "react-hook-form";
 import { ContentType, CreateDispositifRequest, DispositifStatus } from "@refugies-info/api-types";
 import { useEvent } from "hooks";
-import { BaseModal } from "components/Pages/dispositif";
+import BaseModal from "components/UI/BaseModal";
 import { getMissingStepsEdit, Step } from "../functions";
 import CompleteContent from "./CompleteContent";
 import MissingContent from "./MissingContent";
