+++ client/src/components/Pages/dispositif/Edition/MapEdit/DeleteModal/DeleteModal.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Button from "components/UI/Button";
-import { BaseModal } from "components/Pages/dispositif";
+import BaseModal from "components/UI/BaseModal";
 import styles from "./Delete.module.scss";
 
 interface Props {
