+++ client/src/components/Pages/dispositif/Edition/Modals/ModalSponsors/ModalSponsors.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 import { CreateDispositifRequest, Picture, Sponsor } from "@refugies-info/api-types";
-import { BaseModal } from "components/Pages/dispositif";
+import BaseModal from "components/UI/BaseModal";
 import { SimpleFooter, SponsorForm } from "../components";
 import { help } from "./data";
 import styles from "./ModalSponsors.module.scss";
