+++ client/src/components/Pages/dispositif/Edition/Modals/ModalThemes/ModalThemes.tsx
@@ -4,7 +4,7 @@ import { useFormContext } from "react-hook-form";
 import { CreateDispositifRequest } from "@refugies-info/api-types";
 import { themesSelector } from "services/Themes/themes.selectors";
 import EVAIcon from "components/UI/EVAIcon/EVAIcon";
-import { BaseModal } from "components/Pages/dispositif";
+import BaseModal from "components/UI/BaseModal";
 import { SimpleFooter } from "../components";
 import ThemeSelectButton from "./ThemeSelectButton";
 import styles from "./ModalThemes.module.scss";
