+++ client/src/components/Pages/dispositif/StructureReceiveDispositif/StructureReceiveDispositif.tsx
@@ -8,9 +8,9 @@ import { getPath } from "routes";
 import API from "utils/API";
 import { selectedDispositifSelector } from "services/SelectedDispositif/selectedDispositif.selector";
 import Button from "components/UI/Button";
+import BaseModal from "components/UI/BaseModal";
 import { ChoiceButton } from "../Edition";
 import Status from "../Status";
-import BaseModal from "../BaseModal";
 import YesIcon from "assets/dispositif/yes-icon.svg";
 import NoIcon from "assets/dispositif/no-icon.svg";
 import ReceiveDispositif from "assets/dispositif/receive-dispositif.svg";
